requirements.txt
appdirs==1.4.3
certifi==2017.1.23
ChatterBot==0.5.5
click==6.7
future==0.16.0
jsondatabase==0.1.7
nltk==3.2.2
oauthlib==2.0.1
packaging==16.8
pluggy==0.4.0
py==1.4.32
pyparsing==2.2.0
pytest==3.0.6
python-dotenv==0.6.3
python-telegram-bot==5.3.0
python-twitter==3.2.1
requests==2.7.0
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
telebot==0.0.3
urllib3==1.20
virtualenv==15.1.0
